pyproject.toml

[tool.poetry]
packages = [
    { include = "gendiff" }
]

name = "yerkebulanali_gendiff"
version = "0.0.9"
description = "Difference Calculator"
authors = ["Yerkebulan <yerkebulanali@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
PyYAML = "^5.3"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
pytest-cov = "^2.10.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"